+++ app/scripts/services/themes/theme.constants.js
@@ -1,0 +1,13 @@
+/**
+ * Created by C0ZEN on 26/08/2017.
+ */
+(function (angular) {
+    'use strict';
+
+    angular
+        .module('capricieuseApp')
+        .constant('themeConstants', {
+            currentTheme: null
+        });
+
+}(window.angular));